bonus

+++ ft_strlen.c
@@ -10,9 +10,11 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_strlen(char *c)
+#include "libft.h"
+
+size_t		ft_strlen(char *c)
 {
-	int	i;
+	size_t	i;
 	
 	i = 0;
 	while(c[i])
